ChatBot
The program build is an Eliza chatbot that engages in a dialogue with the user. Eliza plays the role of a psychotherapist. Eliza chatbot takes the input from the user and responds with a meaningful message. The chatbox looks for text pattern such as "I feel X" based on the keywords within the input using "word spotting" (such as 'kill', 'no', 'happy', 'angry', 'remember') and responds with a matching comment, such as "I understand. Why do you feel X". An example for this would be if the user enters "I feel sad", then the response would be "Sorry to hear that, tell me more.". The program is robust enough in case it is not able to spot any word, it will return with a default message.

Chatbots were created as an Artificial Intelligence program to chat with users and provide them with information and support instead of having a human being present. Eliza uses a script that is simulated to emulate a psychotherapist.