Add files via upload
This is a question answering system that is developed in Python. The system is only able to answer 'Who', 'What', 'Where' and 'When' questions. The question answering system is not domain specific and provides answers in complete sentences only if user question matches with the system pre defined format. If system does not know the answer to user questions, then it simply returns with default answer that is "I am sorry, I don't know the answer.". And if system is able to find out the output from wikipedia then it returns the result.